requirements.txt
Django==1.11
psycopg2==2.7.1
django-ckeditor==5.2.2